Add fuzz corpus generator from unit tests and integrate into CI
