--- tests/dumps/client_Browse.hex
@@ -1,8 +1,0 @@
-4d 53 47 46 71 00 00 00 01 00 00 00 01 00 00 00  /* MSGFq........... */
-04 00 00 00 04 00 00 00 01 00 0f 02 04 01 00 e3  /* ...............H */
-77 d7 1c 90 95 e2 a4 6d b6 4b e4 0e e0 9e b9 9e  /* ....4~..~.l.q.k. */
-90 25 6a 81 bf d1 01 05 00 00 00 00 00 00 00 ff  /* .%j............. */
-ff ff ff 00 00 00 00 00 00 00 00 00 00 00 00 00  /* ................ */
-00 00 00 00 00 00 00 00 00 00 00 00 01 00 00 00  /* ................ */
-00 55 00 00 00 00 00 00 00 00 00 00 00 3f 00 00  /* .U...........?.. */
-00                                               /* . */
