--- tests/dumps/client_CLO.hex
@@ -1,5 +1,0 @@
-43 4c 4f 46 39 00 00 00 01 00 00 00 01 00 00 00  /* CLOF9........... */
-05 00 00 00 05 00 00 00 01 00 c4 01 04 01 00 e3  /* ...............H */
-77 d7 1c 90 95 e2 a4 6d b6 4b e4 0e e0 9e b9 9e  /* ....4~..~.l.q.k. */
-90 25 6a 81 bf d1 01 05 00 00 00 00 00 00 00 ff  /* .%j............. */
-ff ff ff 00 00 00 00 00 00 00 00                 /* ...........      */
