--- tests/dumps/client_CreateActivateSession.hex
@@ -1,19 +1,0 @@
-4d 53 47 46 72 00 00 00 01 00 00 00 01 00 00 00  /* MSGFr........... */
-02 00 00 00 02 00 00 00 01 00 cd 01 00 00 92 6a  /* ...............j */
-25 6a 81 bf d1 01 02 00 00 00 00 00 00 00 ff ff  /* %j.............. */
-ff ff 10 27 00 00 00 00 00 ff ff ff ff ff ff ff  /* ...'............ */
-ff 00 00 00 00 00 ff ff ff ff ff ff ff ff ff ff  /* ................ */
-ff ff ff ff ff ff ff ff ff ff ff ff ff ff ff ff  /* ................ */
-ff ff ff ff ff ff 00 00 00 00 80 4f 32 41 ff ff  /* ...........O2A.. */
-ff 7f                                            /* .. */
-4d 53 47 46 a1 00 00 00 01 00 00 00 01 00 00 00  /* MSGF............ */
-03 00 00 00 03 00 00 00 01 00 d3 01 04 01 00 e3  /* ...............H */
-77 d7 1c 90 95 e2 a4 6d b6 4b e4 0e e0 9e b9 84  /* ....4~..~.l.q.k. */
-78 25 6a 81 bf d1 01 04 00 00 00 00 00 00 00 ff  /* x%j............. */
-ff ff ff c0 27 09 00 00 00 00 ff ff ff ff ff ff  /* ....'........... */
-ff ff ff ff ff ff ff ff ff ff 01 00 44 01 01 36  /* ............D..6 */
-00 00 00 19 00 00 00 6f 70 65 6e 36 32 35 34 31  /* .......open62541 */
-2d 75 73 65 72 6e 61 6d 65 2d 70 6f 6c 69 63 79  /* -username-policy */
-05 00 00 00 75 73 65 72 31 08 00 00 00 70 61 73  /* ....user1....pas */
-73 77 6f 72 64 ff ff ff ff ff ff ff ff ff ff ff  /* sword........... */
-ff                                               /* . */
