--- tests/dumps/client_HELOPN.hex
@@ -1,13 +1,0 @@
-48 45 4c 46 38 00 00 00 00 00 00 00 00 00 01 00  /* HELF8........... */
-00 00 01 00 00 00 00 01 88 13 00 00 18 00 00 00  /* ................ */
-6f 70 63 2e 74 63 70 3a 2f 2f 6c 6f 63 61 6c 68  /* opc.tcp://localh */
-6f 73 74 3a 34 38 34 32                          /* ost:4842         */
-4f 50 4e 46 85 00 00 00 00 00 00 00 2f 00 00 00  /* OPNF......../... */
-68 74 74 70 3a 2f 2f 6f 70 63 66 6f 75 6e 64 61  /* http://opcfounda */
-74 69 6f 6e 2e 6f 72 67 2f 55 41 2f 53 65 63 75  /* tion.org/UA/Secu */
-72 69 74 79 50 6f 6c 69 63 79 23 4e 6f 6e 65 ff  /* rityPolicy#None. */
-ff ff ff ff ff ff ff 01 00 00 00 01 00 00 00 01  /* .......3........ */
-00 be 01 00 00 97 db 02 02 a8 d2 ce 01 00 00 00  /* ................ */
-00 00 00 00 00 ff ff ff ff 00 00 00 00 00 00 00  /* ................ */
-00 00 00 00 00 00 00 00 01 00 00 00 01 00 00 00  /* ................ */
-00 e0 93 04 00                                   /* .....            */
