--- tests/dumps/client_Write.hex
@@ -1,7 +1,0 @@
-4d 53 47 46 6d 00 00 00 01 00 00 00 01 00 00 00  /* MSGFm........... */
-04 00 00 00 04 00 00 00 01 00 a1 02 04 01 00 e3  /* ...............H */
-77 d7 1c 90 95 e2 a4 6d b6 4b e4 0e e0 9e b9 ba  /* ....4~..~.l.q.k. */
-0f 26 6a 81 bf d1 01 0b 00 00 00 00 00 00 00 ff  /* .&j............. */
-ff ff ff 00 00 00 00 00 00 00 01 00 00 00 03 01  /* ................ */
-00 0a 00 00 00 74 68 65 2e 61 6e 73 77 65 72 0d  /* .....the.answer. */
-00 00 00 ff ff ff ff 01 06 2b 00 00 00           /* .........+...    */
